✨ feat: Implement asynchronous vector store classes for PostgreSQL and MongoDB; enhance document handling and retrieval
