+++ app/services/vector_store/extended_pg_vector.py
@@ -1,0 +1,45 @@
+from typing import Optional
+
+from sqlalchemy import delete
+from sqlalchemy.orm import Session
+from langchain_core.documents import Document
+from langchain_community.vectorstores.pgvector import PGVector
+
+class ExtendedPgVector(PGVector):
+    def get_all_ids(self) -> list[str]:
+        with Session(self._bind) as session:
+            results = session.query(self.EmbeddingStore.custom_id).all()
+            return [result[0] for result in results if result[0] is not None]
+
+    def get_documents_by_ids(self, ids: list[str]) -> list[Document]:
+        with Session(self._bind) as session:
+            results = (
+                session.query(self.EmbeddingStore)
+                .filter(self.EmbeddingStore.custom_id.in_(ids))
+                .all()
+            )
+            return [
+                Document(page_content=result.document, metadata=result.cmetadata or {})
+                for result in results
+                if result.custom_id in ids
+            ]
+
+    def _delete_multiple(
+        self, ids: Optional[list[str]] = None, collection_only: bool = False
+    ) -> None:
+        with Session(self._bind) as session:
+            if ids is not None:
+                self.logger.debug(
+                    "Trying to delete vectors by ids (represented by the model "
+                    "using the custom ids field)"
+                )
+                stmt = delete(self.EmbeddingStore)
+                if collection_only:
+                    collection = self.get_collection(session)
+                    if not collection:
+                        self.logger.warning("Collection not found")
+                        return
+                    stmt = stmt.where(self.EmbeddingStore.collection_id == collection.uuid)
+                stmt = stmt.where(self.EmbeddingStore.custom_id.in_(ids))
+                session.execute(stmt)
+            session.commit()
